feat: implement cache e2e
